Move kernel connection wrapper into api
